requirements.txt
google-api-python-client==2.108.0
google-auth==2.25.2
pdfplumber==0.10.3
PyPDF2==3.0.1
requests==2.31.0
aiohttp==3.9.1
botbuilder-core==4.14.8
botbuilder-schema==4.14.8
